Start handling of const qualification "correctly"
THIS IS AN INTERFACE BREAKING CHANGE: if any const types are
serialized, this will now stop working until the json() methods are
updated. Same for serialization of map and multimap with keys that use
a json() method.

Remove all const_casts. For some reason or other, need to explicitly
specify the version of to_value() used in value_converter<map> and
value_converter<vector>. These are operators::to_value_c() and
operators::to_value_nc().